Refactor e2e kubernetes reporter
- Move to separate package
- Extract common code
- Ensure it fires just after an It() block and before other AfterEach blocks
